refactor: create openai_interface_test.go with interface test
Move TestOpenAIClientImplementsLLMClient to its own file to continue the refactoring of the large openai_client_test.go file. This change:

1. Creates a new file internal/openai/openai_interface_test.go
2. Moves the TestOpenAIClientImplementsLLMClient test to this file
3. Adds a compile-time check for interface compliance
4. Removes the test from the original file
